+++ lib/pages/record/record_model.dart
@@ -4,7 +4,7 @@ import 'package:fl_chart/fl_chart.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:intl/intl.dart';
-import '../const/const.dart';
+import '../../const/const.dart';
 
 class RecordModel {
   final String? _userId = FirebaseAuth.instance.currentUser?.uid;
